services/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
base58==2.1.1
certifi==2025.11.12
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.123.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib-metadata==8.5.0
Markdown==3.7
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
typing-extensions==4.13.2
uvicorn==0.33.0
zipp==3.20.2
